refactor(rest): extract admin audit-schedule routes into routes/admin_audit_schedules.py
Moves the five admin audit-schedule routes (GET/POST audit/schedules,
PATCH/DELETE audit/schedules/{id}, POST .../fire) out of create_app's
closure into routes/admin_audit_schedules.py.

Adds SystemUser, validate_cron and next_fire to routes/deps.py (pure
helpers the closure's _SystemUser/_validate_cron/_next_fire still
duplicate for now — they get collapsed into aliases when the
plugin-job-schedules group, which needs the same three, is extracted
next). The fire-now route calls routes/admin_audit_runs.py's
audit_dispatch directly via RestContext, same as the closure's
_scheduler_fire still does under the old name.
